first commit
Starting from the hw_random driver. Want a sysfs node that we can write a
filesystem path to that will trigger the inotify changed event (to start).

Later need to write a requested event+filename to the node to trigger that
specific inotify event on that specific file.

Studing fsnotify_modify() as the likely fn I need to call to trigger an event.